src/requirements.txt
Django==2.2.7
djangorestframework==3.8.2
django-filter
django-import-export
gunicorn==19.9.0